requirements.txt
asgiref==3.5.2
Django==4.1.1
django-leaflet==0.28.3
djangorestframework==3.13.1
djangorestframework-gis==1.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
